fix types interface
