+++ types/index.ts
@@ -42,3 +42,48 @@ export type SxPropsType = {
   export interface Position {
     coords: Coordinates;
   }
+
+  export interface SideModalCardProps {
+    humidityDegree: string;
+    windDegree: string;
+    isRow: boolean;
+  }
+
+  export interface AlertDialogProps {
+    open: boolean;
+    onClose: () => void;
+    onAgree?: () => void;
+    title: string;
+    content: string;
+    disagreeText?: string;
+    agreeText?: string;
+    deleteColor?: boolean;
+  }
+  export interface AlertDialogTextProps {
+    open: boolean;
+    onClose: () => void;
+    onAgree: (newNote: string) => void;
+    title: string;
+    disagreeText: string;
+    agreeText: string;
+    newNote: string;
+    setNewNote: (note: string) => void;
+  }
+  export interface WeatherModalProps {
+    open: boolean;
+    onClose: () => void;
+    title: string;
+    disagreeText: string;
+    weatherLogo: string;
+    tempDegree: string;
+    humidityDegree: string;
+    windDegree: string;
+    location: string;
+    isError: boolean;
+  }
+
+  export interface WeatherCardProps {
+    time: string;
+    iconSrc: string;
+    temperature: string;
+  }
